now fixed wrapping in mobile
